Add "curpos!" word, to set cursor position.
Uses `int 10h` with AH=2. This is useful for some fun text formatting and display stuff. I feel like it's appropriate for block 31, right next to `curpos@`.

Copies the calling-convention of emit-tty above it. If something's amiss about this, feel free to make changes; I tested it on my end and it worked, but I'm something of an asm noob :)